dropped tga support, back to png
A good TGA decoder/encoder isn't that simple as thought. There isn't
just uncompressed RGB24 and RGB32. RGB565, A1, A8, A16 are part of it
too, all of this can be packed in a colormap with indices size of 1, 8
or 16 bit and furthermore all of the previous can be runlength encoded.
A lot of variations... A TGA V2.0 footer can describe scanline length
and gamma correction. Much work to fulfill all user wishes.